migrate to e2e test
