+++ test/development/middleware-errors/index.test.ts
@@ -1,0 +1,344 @@
+import {
+  check,
+  fetchViaHTTP,
+  File,
+  findPort,
+  getRedboxSource,
+  hasRedbox,
+  killApp,
+  launchApp,
+} from 'next-test-utils'
+import { join } from 'path'
+import stripAnsi from 'strip-ansi'
+import webdriver from 'next-webdriver'
+import { createNextDescribe } from 'e2e-utils'
+
+// const context = {
+//   appDir: join(__dirname, '../'),
+//   buildLogs: { output: '', stdout: '', stderr: '' },
+//   logs: { output: '', stdout: '', stderr: '' },
+//   middleware: new File(join(__dirname, '../middleware.js')),
+//   page: new File(join(__dirname, '../pages/index.js')),
+// }
+
+createNextDescribe(
+  'middleware - development errors',
+  {
+    files: __dirname,
+    env: { __NEXT_TEST_WITH_DEVTOOL: '1' },
+  },
+  ({ next }) => {
+    beforeEach(async () => {
+      await next.stop()
+    })
+    // beforeEach(async () => {
+    //   context.logs = { output: '', stdout: '', stderr: '' }
+    //   context.appPort = await findPort()
+    //   context.app = await launchApp(context.appDir, context.appPort, {
+    //     env: { __NEXT_TEST_WITH_DEVTOOL: 1 },
+    //     stderr: false,
+    //     stdout: false,
+    //     onStdout(msg) {
+    //       // console.trace('onStdout', msg)
+    //       next.cliOutput += msg
+    //       context.logs.stdout += msg
+    //     },
+    //     onStderr(msg) {
+    //       // console.trace('onStderr', msg)
+    //       next.cliOutput += msg
+    //       context.logs.stderr += msg
+    //     },
+    //   })
+    // })
+
+    // afterEach(async () => {
+    //   context.middleware.restore()
+    //   context.page.restore()
+    //   if (context.app) {
+    //     await killApp(context.app)
+    //   }
+    // })
+
+    describe('when middleware throws synchronously', () => {
+      beforeEach(async () => {
+        await next.patchFile(
+          'middleware.js',
+          `
+      export default function () {
+        throw new Error('boom')
+      }`
+        )
+
+        await next.start()
+      })
+
+      it('logs the error correctly', async () => {
+        await next.fetch('/')
+        const output = stripAnsi(next.cliOutput)
+        await check(
+          () => stripAnsi(next.cliOutput),
+          new RegExp(
+            `middleware.js \\(\\d+:\\d+\\) @ Object.default \\[as handler\\]\\s*\\w+\\s*boom`,
+            'm'
+          )
+        )
+        expect(output).not.toContain(
+          'webpack-internal:///(middleware)/./middleware.js'
+        )
+      })
+
+      it('renders the error correctly and recovers', async () => {
+        const browser = await next.browser('/')
+        expect(await hasRedbox(browser, true)).toBe(true)
+        await next.patchFile('middleware.js', `export default function () {}`)
+        await hasRedbox(browser, false)
+      })
+    })
+
+    describe('when middleware contains an unhandled rejection', () => {
+      beforeEach(async () => {
+        await next.patchFile(
+          'middleware.js',
+          `
+      import { NextResponse } from 'next/server'
+      async function throwError() {
+        throw new Error('async boom!')
+      }
+      export default function () {
+        throwError()
+        return NextResponse.next()
+      }`
+        )
+
+        await next.start()
+      })
+
+      it('logs the error correctly', async () => {
+        await next.fetch('/')
+        await check(
+          () => stripAnsi(next.cliOutput),
+          new RegExp(`unhandledRejection: Error: async boom!`, 'm')
+        )
+        // expect(output).not.toContain(
+        //   'webpack-internal:///(middleware)/./middleware.js'
+        // )
+      })
+
+      it('does not render the error', async () => {
+        const browser = await next.browser('/')
+        expect(await hasRedbox(browser, false)).toBe(false)
+        expect(await browser.elementByCss('#page-title')).toBeTruthy()
+      })
+    })
+
+    describe('when running invalid dynamic code with eval', () => {
+      beforeEach(async () => {
+        await next.patchFile(
+          'middleware.js',
+          `
+      import { NextResponse } from 'next/server'
+      export default function () {
+        eval('test')
+        return NextResponse.next()
+      }`
+        )
+
+        await next.start()
+      })
+
+      it('logs the error correctly', async () => {
+        await next.fetch('/')
+        const output = stripAnsi(next.cliOutput)
+        await check(
+          () => stripAnsi(next.cliOutput),
+          new RegExp(
+            `middleware.js \\(\\d+:\\d+\\) @ eval\\s*\\w+\\s*test is not defined`,
+            'm'
+          )
+        )
+        expect(output).not.toContain(
+          'webpack-internal:///(middleware)/./middleware.js'
+        )
+      })
+
+      it('renders the error correctly and recovers', async () => {
+        const browser = await next.browser('/')
+        expect(await hasRedbox(browser, true)).toBe(true)
+        expect(await getRedboxSource(browser)).toContain(`eval('test')`)
+        await next.patchFile('middleware.js', `export default function () {}`)
+        await hasRedbox(browser, false)
+      })
+    })
+
+    describe('when throwing while loading the module', () => {
+      beforeEach(async () => {
+        await next.patchFile(
+          'middleware.js',
+          `
+      import { NextResponse } from 'next/server'
+      throw new Error('booooom!')
+      export default function () {
+        return NextResponse.next()
+      }`
+        )
+        await next.start()
+      })
+
+      it('logs the error correctly', async () => {
+        await next.fetch('/')
+        const output = stripAnsi(next.cliOutput)
+        await check(
+          () => stripAnsi(next.cliOutput),
+          new RegExp(
+            `middleware.js \\(\\d+:\\d+\\) @ <unknown>\\s*\\w+\\s*booooom!`,
+            'm'
+          )
+        )
+        expect(output).not.toContain(
+          'webpack-internal:///(middleware)/./middleware.js'
+        )
+      })
+
+      it('renders the error correctly and recovers', async () => {
+        const browser = await next.browser('/')
+        expect(await hasRedbox(browser, true)).toBe(true)
+        expect(await getRedboxSource(browser)).toContain(
+          `throw new Error('booooom!')`
+        )
+        await next.patchFile('middleware.js', `export default function () {}`)
+        await hasRedbox(browser, false)
+      })
+    })
+
+    describe('when there is an unhandled rejection while loading the module', () => {
+      beforeEach(async () => {
+        await next.patchFile(
+          'middleware.js',
+          `
+      import { NextResponse } from 'next/server'
+      (async function(){
+        throw new Error('you shall see me')
+      })()
+
+      export default function () {
+        return NextResponse.next()
+      }`
+        )
+
+        await next.start()
+      })
+
+      it('logs the error correctly', async () => {
+        await next.fetch('/')
+        await check(
+          () => stripAnsi(next.cliOutput),
+          new RegExp(`unhandledRejection: Error: you shall see me`, 'm')
+        )
+        // expect(output).not.toContain(
+        //   'webpack-internal:///(middleware)/./middleware.js'
+        // )
+      })
+
+      it('does not render the error', async () => {
+        const browser = await next.browser('/')
+        expect(await hasRedbox(browser, false)).toBe(false)
+        expect(await browser.elementByCss('#page-title')).toBeTruthy()
+      })
+    })
+
+    describe('when there is an unhandled rejection while loading a dependency', () => {
+      beforeEach(async () => {
+        await next.patchFile(
+          'middleware.js',
+          `
+      import { NextResponse } from 'next/server'
+      import './lib/unhandled'
+
+      export default function () {
+        return NextResponse.next()
+      }`
+        )
+
+        await next.start()
+      })
+
+      it('logs the error correctly', async () => {
+        await next.fetch('/')
+        const output = stripAnsi(next.cliOutput)
+        await check(
+          () => stripAnsi(next.cliOutput),
+          new RegExp(
+            ` uncaughtException: Error: This file asynchronously fails while loading`,
+            'm'
+          )
+        )
+        expect(output).not.toContain(
+          'webpack-internal:///(middleware)/./middleware.js'
+        )
+      })
+
+      it('does not render the error', async () => {
+        const browser = await next.browser('/')
+        expect(await hasRedbox(browser, false)).toBe(false)
+        expect(await browser.elementByCss('#page-title')).toBeTruthy()
+      })
+    })
+
+    describe('when there is a compilation error from boot', () => {
+      beforeEach(async () => {
+        await next.patchFile('middleware.js', `export default function () }`)
+
+        await next.start()
+      })
+
+      it('logs the error correctly', async () => {
+        await next.fetch('/')
+        expect(next.cliOutput).toContain(`Expected '{', got '}'`)
+        expect(next.cliOutput.split(`Expected '{', got '}'`).length).toEqual(2)
+      })
+
+      it('renders the error correctly and recovers', async () => {
+        const browser = await next.browser('/')
+        expect(await hasRedbox(browser, true)).toBe(true)
+        expect(
+          await browser
+            .elementByCss('#nextjs__container_build_error_label')
+            .text()
+        ).toEqual('Failed to compile')
+        await next.patchFile('middleware.js', `export default function () {}`)
+        await hasRedbox(browser, false)
+        expect(await browser.elementByCss('#page-title')).toBeTruthy()
+      })
+    })
+
+    describe('when there is a compilation error after boot', () => {
+      beforeEach(async () => {
+        await next.patchFile('middleware.js', `export default function () {}`)
+
+        await next.start()
+      })
+
+      it('logs the error correctly', async () => {
+        await next.fetch('/')
+        await next.patchFile('middleware.js', `export default function () }`)
+        await check(() => {
+          expect(next.cliOutput).toContain(`Expected '{', got '}'`)
+          expect(next.cliOutput.split(`Expected '{', got '}'`).length).toEqual(
+            2
+          )
+          return 'success'
+        }, 'success')
+      })
+
+      it('renders the error correctly and recovers', async () => {
+        const browser = await next.browser('/')
+        expect(await hasRedbox(browser, false)).toBe(false)
+        await next.patchFile('middleware.js', `export default function () }`)
+        expect(await hasRedbox(browser, true)).toBe(true)
+        await next.patchFile('middleware.js', `export default function () {}`)
+        expect(await hasRedbox(browser, false)).toBe(false)
+        expect(await browser.elementByCss('#page-title')).toBeTruthy()
+      })
+    })
+  }
+)
