--- test/integration/middleware-dev-errors/middleware.js
@@ -1,1 +1,0 @@
-// this will be populated by each test
